refactor: rework vue map-components template to match cli
