+++ packages/map-components/templates/vue/.vscode/extensions.json
@@ -1,0 +1,3 @@
+{
+  "recommendations": ["Vue.volar", "Vue.vscode-typescript-vue-plugin"]
+}
