+++ packages/map-components/templates/vue/src/assets/main.css
